Wave Management Base Finish
News :

Added a system of gestion off enemy in Wave

When the wave have 0 enemy, we start the next, later the shop session.

Event Added to the gestion of the enemies

--> Maybe use it for the autoTarget !

Next Upgrade : Add a limit of monster, an intern timer in WaveManager to control all the lag caused by entities.